Move deployment related files to deploy folder (#1086)
* Moved deployment related files
