+++ src/commands/createWebApp/setAppWizardContextDefault.ts
@@ -5,17 +5,12 @@
 
 import * as fse from 'fs-extra';
 import * as path from 'path';
-import { ConfigurationTarget, workspace, WorkspaceFolder } from 'vscode';
+import { workspace, WorkspaceFolder } from 'vscode';
 import { IAppServiceWizardContext, LinuxRuntimes, WebsiteOS } from 'vscode-azureappservice';
-import { IActionContext, ICreateChildImplContext, LocationListStep } from 'vscode-azureextensionui';
+import { ICreateChildImplContext, LocationListStep } from 'vscode-azureextensionui';
 import { javaUtils } from '../../utils/javaUtils';
 import { findFilesByFileExtension, getContainingWorkspace } from '../../utils/workspace';
-
-export interface IDeployWizardContext extends IActionContext {
-    fsPath?: string;
-    deployedWithConfigs?: boolean;
-    configurationTarget?: ConfigurationTarget;
-}
+import { IDeployWizardContext } from '../deploy/IDeployWizardContext';
 
 export async function setAppWizardContextDefault(wizardContext: IAppServiceWizardContext & Partial<IDeployWizardContext> & Partial<ICreateChildImplContext>): Promise<void> {
     // if the user entered through "Deploy", we'll have a project to base our recommendations on
