+++ src/explorer/WebAppTreeItem.ts
@@ -10,7 +10,7 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import { ISiteTreeRoot } from 'vscode-azureappservice';
 import { AzExtTreeItem, AzureTreeItem, createAzureClient, IActionContext, openInPortal } from 'vscode-azureextensionui';
-import { deploy } from '../commands/deploy';
+import { deploy } from '../commands/deploy/deploy';
 import { AppServiceDialogResponses, extensionPrefix } from '../constants';
 import { ext } from '../extensionVariables';
 import { nonNullProp, nonNullValue } from '../utils/nonNull';
