+++ src/extension.ts
@@ -18,7 +18,7 @@ import { revealConnection } from './commands/connections/revealConnection';
 import { revealConnectionInAppSettings } from './commands/connections/revealConnectionInAppSettings';
 import { createSlot } from './commands/createSlot';
 import { createWebApp, createWebAppAdvanced } from './commands/createWebApp/createWebApp';
-import { deploy } from './commands/deploy';
+import { deploy } from './commands/deploy/deploy';
 import { connectToGitHub } from './commands/deployments/connectToGitHub';
 import { disconnectRepo } from './commands/deployments/disconnectRepo';
 import { editScmType } from './commands/deployments/editScmType';
